+++ index.js
@@ -41,6 +41,7 @@ const {getChannelBalance} = require('./lnd_methods');
 const {getChannels} = require('./lnd_methods');
 const {getClosedChannels} = require('./lnd_methods');
 const {getConnectedWatchtowers} = require('./lnd_methods');
+const {getEphemeralChannelIds} = require('./lnd_methods');
 const {getFailedPayments} = require('./lnd_methods');
 const {getFeeRates} = require('./lnd_methods');
 const {getForwardingConfidence} = require('./lnd_methods');
@@ -183,11 +184,12 @@ module.exports = {
   getChainFeeEstimate,
   getChainFeeRate,
   getChainTransactions,
-  getChannelBalance,
   getChannel,
+  getChannelBalance,
   getChannels,
   getClosedChannels,
   getConnectedWatchtowers,
+  getEphemeralChannelIds,
   getFailedPayments,
   getFeeRates,
   getForwardingConfidence,