+++ lnd_methods/index.js
@@ -39,6 +39,7 @@ const {getChannelBalance} = require('./offchain');
 const {getChannels} = require('./offchain');
 const {getClosedChannels} = require('./offchain');
 const {getConnectedWatchtowers} = require('./offchain');
+const {getEphemeralChannelIds} = require('./offchain');
 const {getFailedPayments} = require('./offchain');
 const {getFeeRates} = require('./offchain');
 const {getForwardingConfidence} = require('./offchain');
@@ -182,6 +183,7 @@ module.exports = {
   getChannels,
   getClosedChannels,
   getConnectedWatchtowers,
+  getEphemeralChannelIds,
   getFailedPayments,
   getFeeRates,
   getForwardingConfidence,