+++ lnd_methods/offchain/index.js
@@ -15,6 +15,7 @@ const getChannelBalance = require('./get_channel_balance');
 const getChannels = require('./get_channels');
 const getClosedChannels = require('./get_closed_channels');
 const getConnectedWatchtowers = require('./get_connected_watchtowers');
+const getEphemeralChannelIds = require('./get_ephemeral_channel_ids');
 const getFailedPayments = require('./get_failed_payments');
 const getFeeRates = require('./get_fee_rates');
 const getForwardingConfidence = require('./get_forwarding_confidence');
@@ -71,6 +72,7 @@ module.exports = {
   getChannels,
   getClosedChannels,
   getConnectedWatchtowers,
+  getEphemeralChannelIds,
   getFailedPayments,
   getFeeRates,
   getForwardingConfidence,