+++ test/lnd_responses/test_channel_accept_as_open_request.js
@@ -35,6 +35,7 @@ const tests = [
       csv_delay: 1,
       id: Buffer.alloc(32).toString('hex'),
       is_private: true,
+      is_trusted_funding: false,
       local_balance: 1,
       local_reserve: 1,
       max_pending_mtokens: '1000',